requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
aws-sam-translator==1.97.0
aws-xray-sdk==2.14.0
black==25.1.0
boto3==1.38.13
botocore==1.38.13
certifi==2025.4.26
cffi==1.17.1
cfn-lint==1.35.3
charset-normalizer==3.4.2
click==8.2.0
cryptography==45.0.2
dnspython==2.7.0
docker==7.1.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
graphql-core==3.2.6
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
joserfc==1.0.4
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moto==5.1.4
mpmath==1.3.0
multipart==1.2.1
mypy_extensions==1.1.0
networkx==3.4.2
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
packaging==25.0
pathable==0.4.4
pathspec==0.12.1
pg8000==1.31.2
platformdirs==4.3.8
pluggy==1.5.0
ply==3.11
py-partiql-parser==0.6.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
pytest-testdox==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
responses==0.25.7
rfc3339-validator==0.1.4
rich==14.0.0
rich-toolkit==0.14.6
rpds-py==0.25.1
s3transfer==0.12.0
scramp==1.4.5
setuptools==80.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xmltodict==0.14.2
